cleanup: rename filename for consistency
